Comprehensive clippy cleanup and panic fixes
- Fixed critical panic in burn_iris_loader.rs with proper error handling
- Eliminated ALL clippy warnings (40+ reduced to 0)
- Enhanced error handling throughout codebase
- Fixed needless_range_loop, single_range_in_vec_init, and other style warnings
- Improved code quality with idiomatic Rust patterns
- All tests passing (16/16)
- Clean compilation across all targets
- Added comprehensive brain dynamics examples and documentation

Status: Production-ready codebase with zero static analysis warnings

+++ examples/sine_wave_training.rs
@@ -120,7 +120,7 @@ fn test_model_predictions(model: &FullLinossModel<MyBackend>, device: &<MyBacken
 
     let mse: f32 = test_targets.iter().zip(pred_data.iter())
         .map(|(t, p)| {
-            let diff = *t as f32 - *p as f32;
+            let diff = *t - *p;
             diff.powi(2)
         })
         .sum::<f32>() / test_targets.len() as f32;
